# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Budget API",
        "description": "API endpoints for managing budgets",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/works": {
            "get": {
                "tags": [
                    "Works"
                ],
                "summary": "Get all works",
                "operationId": "3a1370667b03c67db44e4696290559df",
                "responses": {
                    "200": {
                        "description": "List of works retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Work"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Works"
                ],
                "summary": "Create a new work",
                "operationId": "06f41cbe7a7c7bf853dd37caad3e72ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Work created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/works/{id}": {
            "get": {
                "tags": [
                    "Works"
                ],
                "summary": "Get work by ID",
                "operationId": "3037e3afce3b7bb9b2ccbf59139f26f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Work ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Work found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Work not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Works"
                ],
                "summary": "Update an existing work",
                "operationId": "0851a69a4efa18b7682fd6bcb14147ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Work updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Works"
                ],
                "summary": "Delete a work",
                "operationId": "02bd1c51ebfb34a53c0384eae62044b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Work deleted successfully"
                    },
                    "404": {
                        "description": "Work not found"
                    }
                }
            }
        },
        "/api/v1/works/materials/{id}": {
            "post": {
                "tags": [
                    "Works"
                ],
                "summary": "Add materials to a work",
                "operationId": "e34d5623eba9faf2abb738ef69e07b18",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Work ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddMaterialsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Materials added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/works/states/{id}/{state}": {
            "post": {
                "tags": [
                    "Works"
                ],
                "summary": "Change work state",
                "operationId": "8086666a35ca1fb26b75dc9fa78c1365",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Work ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "New state",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "State changed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AddMaterialsRequest": {
                "required": [
                    "invoice_id",
                    "materials"
                ],
                "properties": {
                    "invoice_id": {
                        "type": "integer"
                    },
                    "materials": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "quantity": {
                                    "type": "number"
                                },
                                "price": {
                                    "type": "number"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "Work": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "order": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "estimated_time": {
                        "type": "integer"
                    },
                    "dead_line": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "budget_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "WorkRequest": {
                "required": [
                    "order",
                    "name",
                    "estimated_time",
                    "dead_line",
                    "budget_id"
                ],
                "properties": {
                    "order": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "estimated_time": {
                        "type": "integer"
                    },
                    "dead_line": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "budget_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}